+++ src/jtl/Connector/Shopware/Controller/Customer.php
@@ -59,10 +59,12 @@ class Customer extends DataController
                     
                     // Birthday
                     try {
-                        if (isset($customerSW['birthday']) && strlen($customerSW['birthday']) > 0
+                        if (isset($customerSW['birthday']) && !($customerSW['birthday'] instanceof \DateTime)
+                            && strlen($customerSW['birthday']) > 0
                             && $customerSW['birthday'] !== '0000-00-00') {
                             $customer->setBirthday(new \DateTime($customerSW['birthday']));
-                        } elseif (isset($customerSW['birthday']) && $customerSW['birthday'] instanceof \DateTime) {
+                        } elseif (isset($customerSW['birthday']) && $customerSW['birthday'] instanceof \DateTime
+                            && $customerSW['birthday'] != new \DateTime('0000-00-00')) {
                             $customer->setBirthday($customerSW['birthday']);
                         }
                     } catch (\Exception $e) { }