+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+2.0.16
+------
+- Fixed customer birthday property warning
+- Added customer birthday property wawi exception workaround
+- Added customer order pull start date to payments
+- Added product attribute sw_image_config_ignores to configure product image assignments
+- Added product image titles via wawi image alt text
+- Removed deprecated method call
+
 2.0.15
 ------
 - Fixed customer birthday and title